requirement.txt
aiogram==3.1.1
playwright==1.40.0
sqlalchemy==2.0.23
cryptography==4.1.5
python-dotenv==1.0.0
